[API] rename schooling-registration by organization-learner in validators
